BookMacster 1.21 (was 1.20.7), prior to Code Review.   After updating to Xcode 5.1 which wanted lots of stupid integer typecasts, finally fixed the #import Sparkle/SUUpdater.h phenomena (removed the path), added “Periodic” advanced trigger.

+++ NSData+HexString.m
@@ -33,7 +33,7 @@ char ASCIIHexCharacterForNibble(short nibble, BOOL uppercaseLetters) {
 		NSData* dataChunk = [self subdataWithRange:NSMakeRange(i, 1)] ;
 		unsigned char oneByte ;
 		[dataChunk getBytes:&oneByte] ;
-		int subscript ;
+		NSInteger subscript ;
 		subscript = 2*i ;
 		hashCString[subscript] = ASCIIHexCharacterForNibble((oneByte & 0xf0) >> 4, NO) ;
 		subscript = 2*i+1 ;
